07-07-23 11:20
Most up to date version of Models.ipynb.

SKLearn Prep now performs a manual KFold on the features/labels arrays to then be iterated through when F.E. respective section per classifier.

SVM and Decision Trees can run. Decision Trees gives 100% accuracy and Recall, which is weird. Categorical Naive-bayes throws an error, which should be listed in the ### Notes in that section.